Consolidates Google and Azure providers
